requirements.txt
EdgeGPT==0.8.1
PySide6==6.5.0
PySide6_Addons==6.5.0
PySide6_Essentials==6.5.0
qasync
tiktoken
bs4
python-pptx
pdfminer.six
docx2txt
aiohttp
openai
Pillow